fixed ui lib imports, moved stories to comp dir
