+++ packages/ui/package.json
@@ -2,6 +2,8 @@
   "name": "@solvedcard/ui",
   "version": "0.0.1",
   "private": true,
+  "main": "lib/src/index.js",
+  "types": "lib/src/index.d.ts",
   "dependencies": {
     "@craco/craco": "^6.1.1",
     "@testing-library/jest-dom": "^5.11.4",
@@ -14,6 +16,7 @@
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-scripts": "4.0.3",
+    "timeago.js": "^4.0.2",
     "typescript": "^4.1.2",
     "web-vitals": "^1.0.1"
   },
