+++ packages/ui/src/components/Button/Button.stories.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 // also exported from '@storybook/react' if you can deal with breaking changes in 6.1
 import { Story, Meta } from "@storybook/react/types-6-0";
 
-import { Button, ButtonProps } from "../components/Button/Button";
+import { Button, ButtonProps } from "./Button";
 
 export default {
 	title: "Example/Button",
