+++ packages/ui/src/components/DataTable/DataTable.stories.tsx
@@ -2,8 +2,8 @@ import * as React from "react";
 // also exported from '@storybook/react' if you can deal with breaking changes in 6.1
 import { Story, Meta } from "@storybook/react/types-6-0";
 
-import { DataTable } from "../components/DataTable/DataTable";
-import { DataTableProps } from "../components/DataTable/types";
+import { DataTable } from "./DataTable";
+import { DataTableProps } from "./types";
 
 export default {
 	title: "DataTable",
